% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utils.R
\name{city_outfit}
\alias{city_outfit}
\title{Recommend by City Name (with State detection)}
\usage{
city_outfit(city)
}
\arguments{
\item{city}{A character string representing the city name (e.g., "Tampa, FL", "Tampa").}
}
\description{
Recommend by City Name (with State detection)
}
