% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_utils.R
\name{coords_outfit}
\alias{coords_outfit}
\title{Recommend by Coordinates}
\usage{
coords_outfit(lat, long)
}
\arguments{
\item{lat}{Numeric. The latitude of the target node.}

\item{long}{Numeric. The longitude of the target node.}
}
\description{
Recommend by Coordinates
}
