% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CShell_capitals}
\alias{CShell_capitals}
\title{US City Capitals Dataset}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 50 rows and 16 columns.
}
\usage{
CShell_capitals
}
\description{
A reference table for tactical deployment to state capitals.
}
\keyword{datasets}
